pyproject.toml
[tool.poetry]
name = "support"
version = "0.1.0"
description = ""
authors = ["ermakoventer <ermakoventer777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.2"
celery = "^5.2.7"
redis = "^4.3.4"
psycopg-binary = "^3.1.3"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.1"
psycopg2 = "^2.9.4"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
